+++ src/drivers/touch/port/esp_lcd_touch_cst820.h
@@ -18,12 +18,12 @@
 extern "C" {
 #endif
 
-#define ESP_LCD_TOUCH_CST816S_VER_MAJOR    (0)
-#define ESP_LCD_TOUCH_CST816S_VER_MINOR    (0)
-#define ESP_LCD_TOUCH_CST816S_VER_PATCH    (1)
+#define ESP_LCD_TOUCH_CST820_VER_MAJOR    (0)
+#define ESP_LCD_TOUCH_CST820_VER_MINOR    (1)
+#define ESP_LCD_TOUCH_CST820_VER_PATCH    (0)
 
 /**
- * @brief Create a new CST816S touch driver
+ * @brief Create a new CST820 touch driver
  *
  * @note  The I2C communication should be initialized before use this function.
  *
@@ -36,18 +36,18 @@ extern "C" {
 esp_err_t esp_lcd_touch_new_i2c_cst820(const esp_lcd_panel_io_handle_t io, const esp_lcd_touch_config_t *config, esp_lcd_touch_handle_t *tp);
 
 /**
- * @brief I2C address of the CST816S controller
+ * @brief I2C address of the CST820 controller
  *
  */
-#define ESP_LCD_TOUCH_IO_I2C_CST816S_ADDRESS    (0x15)
+#define ESP_LCD_TOUCH_IO_I2C_CST820_ADDRESS    (0x15)
 
 /**
  * @brief Touch IO configuration structure
  *
  */
-#define ESP_LCD_TOUCH_IO_I2C_CST816S_CONFIG()             \
+#define ESP_LCD_TOUCH_IO_I2C_CST820_CONFIG()             \
     {                                                    \
-        .dev_addr = ESP_LCD_TOUCH_IO_I2C_CST816S_ADDRESS, \
+        .dev_addr = ESP_LCD_TOUCH_IO_I2C_CST820_ADDRESS, \
         .control_phase_bytes = 1,                        \
         .dc_bit_offset = 0,                              \
         .lcd_cmd_bits = 8,                              \