+++ src/esp_panel_versions.h
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2024 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2024-2025 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -12,15 +12,15 @@
 
 /* File `esp_panel_drivers_conf.h` */
 #define ESP_PANEL_DRIVERS_CONF_VERSION_MAJOR 1
-#define ESP_PANEL_DRIVERS_CONF_VERSION_MINOR 0
+#define ESP_PANEL_DRIVERS_CONF_VERSION_MINOR 1
 #define ESP_PANEL_DRIVERS_CONF_VERSION_PATCH 0
 
 /* File `esp_panel_board_custom_conf.h` */
 #define ESP_PANEL_BOARD_CUSTOM_VERSION_MAJOR 1
-#define ESP_PANEL_BOARD_CUSTOM_VERSION_MINOR 1
+#define ESP_PANEL_BOARD_CUSTOM_VERSION_MINOR 2
 #define ESP_PANEL_BOARD_CUSTOM_VERSION_PATCH 0
 
 /* File `esp_panel_board_supported_conf.h` */
 #define ESP_PANEL_BOARD_SUPPORTED_VERSION_MAJOR 1
-#define ESP_PANEL_BOARD_SUPPORTED_VERSION_MINOR 0
+#define ESP_PANEL_BOARD_SUPPORTED_VERSION_MINOR 1
 #define ESP_PANEL_BOARD_SUPPORTED_VERSION_PATCH 0