+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * feat(docs): provides an example of how to integrate this library into micropython by @tsteinruecken (#190)
 * feat(board): add board Waveshare:ESP32_S3_TOUCH_LCD_1_85_C @hegdi (#208)
+* feat(board): add some new VIEWE boards @VIEWESMART (#197)
+* feat(drivers): add touch controller CST820 @VIEWESMART (#197)
 
 ### Bugfixes: